# packages/backend/public/swagger.json
{
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "info": {
        "version": "0.0.1",
        "title": "renkei-backend",
        "description": "",
        "license": {
            "name": "UNLICENSED"
        }
    },
    "paths": {
...
        "/auth/challenge/{address}": {
            "get": {
                "summary": "Retrieve a challenge for a given user",
                "description": "Retrieve a challenge for a given user.\nThis challenge is used to sign a message that will be used to authenticate the user.",
                "tags": ["auth"],
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A promise that resolves to the challenge response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetChallengeResponseDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/challenge": {
            "post": {
                "summary": "Create a new challenge",
                "description": "Create a new challenge.",
                "tags": ["auth"],
                "parameters": [],
                "requestBody": {
                    "description": "- The body parameters for creating the challenge.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateChallengeBodyDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "A promise that resolves to the challenge creation response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateChallengeResponseDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/token": {
            "post": {
                "summary": "Create a new token",
                "description": "Create a new token.",
                "tags": ["auth"],
                "parameters": [],
                "requestBody": {
                    "description": "- The body parameters for creating the token.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTokenBodyDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "A promise that resolves to the token creation response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateTokenResponseDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "summary": "Refresh an existing token",
                "description": "Refresh an existing token.",
                "tags": ["auth"],
                "parameters": [],
                "requestBody": {
                    "description": "- The body parameters for refreshing the token.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshTokenBodyDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "A promise that resolves to the token refresh response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RefreshTokenResponseDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GetChallengeResponseDto": {
                "type": "object",
                "properties": {
                    "challenge": {
                        "type": "string",
                        "title": "The challenge string",
                        "description": "The challenge string."
                    }
                },
                "required": ["challenge"],
                "description": "Data Transfer Object for GetChallengeResponse.\nThis class is used to encapsulate the response data for a challenge request."
            },
            "CreateChallengeResponseDto": {
                "type": "object",
                "properties": {
                    "challenge": {
                        "type": "string",
                        "title": "The challenge string",
                        "description": "The challenge string."
                    }
                },
                "required": ["challenge"],
                "description": "Represents the response DTO for creating a challenge."
            },
            "CreateChallengeBodyDto": {
                "type": "object",
                "properties": {
                    "address": {
                        "type": "string",
                        "title": "The address associated with the challenge",
                        "description": "The address associated with the challenge."
                    }
                },
                "required": ["address"],
                "description": "Data Transfer Object for creating a challenge."
            },
            "CreateTokenResponseDto": {
                "type": "object",
                "properties": {
                    "accessToken": {
                        "type": "string",
                        "title": "The access token string",
                        "description": "The access token string."
                    },
                    "refreshToken": {
                        "type": "string",
                        "title": "The refresh token string",
                        "description": "The refresh token string."
                    }
                },
                "required": ["accessToken", "refreshToken"],
                "description": "Data Transfer Object (DTO) for creating a token response."
            },
            "CreateTokenBodyDto": {
                "type": "object",
                "properties": {
                    "address": {
                        "type": "string",
                        "title": "The address associated with the token",
                        "description": "The address associated with the token."
                    },
                    "signature": {
                        "type": "string",
                        "title": "The signature associated with the token",
                        "description": "The signature associated with the token."
                    }
                },
                "required": ["address", "signature"],
                "description": "Data Transfer Object for creating a token."
            },
            "RefreshTokenResponseDto": {
                "type": "object",
                "properties": {
                    "accessToken": {
                        "type": "string",
                        "title": "The access token string",
                        "description": "The access token string."
                    },
                    "refreshToken": {
                        "type": "string",
                        "title": "The refresh token string",
                        "description": "The refresh token string."
                    }
                },
                "required": ["accessToken", "refreshToken"],
                "description": "Data Transfer Object (DTO) for the response containing the refreshed tokens."
            },
            "RefreshTokenBodyDto": {
                "type": "object",
                "properties": {
                    "refreshToken": {
                        "type": "string",
                        "title": "The refresh token string",
                        "description": "The refresh token string."
                    }
                },
                "required": ["refreshToken"],
                "description": "Data Transfer Object (DTO) for handling refresh token requests."
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}